:sparkles: Tweak how escape works some more
In this case, if focus is down in an activity, draw focus back up to the
buttons. At this point I'm starting to think I should stop overloading the
binding at the screen level and, instead, have a binding at the types
buttons level, at the activity level, and at the filter panel level.

But this gets the flow I want going.