:hammer: Improve how we process the escape key
While I want to carry on overloading the use of the escape key, I didn't
like all the logic in the app, and I also didn't like how the title in the
footer couldn't ever take on the correct context. So here I move different
bindings into the relevant widgets and process things where it makes sense.