:hammer: Switch from one monolithic script to multiple files
The application started out pretty small, just living in one script, but
pretty quickly it got kinda hefty. I'm not really much of a fan of
monolithic scripts so it felt like now was a good time to split this up --
especially given I want to move to actually deploying it to PyPi so others
can easily try it out.

Also... now felt like a good time to start playing with splitting up a
Textual application and also this was going to be a good test of me leaning
on using messages to make calls that are child->parent in nature.