:bug: Fix main display not refreshing correctly when adding new item
This didn't seem to be a problem with earlier versions of Textual, but as of
the upgrade to 0.14 I was seeing the Vertical that holds the activities not
updating other than the first item in the list -- I had to move the mouse or
something to get it to redraw.

This fixes that (and, in retrospect, makes a lot more sense too).